Add swift-playgrounds sample file
The Swift Playgrounds Book for micro:bit operates in Bluetooth
tethered mode. This example is used for the C++ build of the hex
file for using Swift Playgrounds with the device.

It's necessary for the DFU service to exist, because the Swift
Playgrounds Book uses this to determine that it can/should pair
with the micro:bit

Tested with with Swift 4.1 version of the playground book. V1.2.0